Create a separate `color_scale` module
