+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod ansi;
 pub mod cmyk;
+pub mod color_scale;
 pub mod colorspace;
 pub mod convert;
 pub mod delta_e;
@@ -24,7 +25,6 @@ mod test_helper;
 
 use std::{fmt, str::FromStr};
 
-// Re-export color space types
 pub use cmyk::CMYK;
 pub use hsl::HSLA;
 pub use hsv::HSVA;
@@ -39,6 +39,8 @@ use colorspace::ColorSpace;
 pub use helper::Fraction;
 use helper::MaxPrecision;
 use types::{Hue, Scalar};
+
+pub use color_scale::ColorScale;
 
 /// The representation of a color.
 ///
@@ -682,96 +684,6 @@ pub enum Format {
     NoSpaces,
 }
 
-/// The representation of a color stop for a `ColorScale`.
-/// The position defines where the color is placed from left (0.0) to right (1.0).
-#[derive(Debug, Clone)]
-struct ColorStop {
-    color: Color,
-    position: Fraction,
-}
-
-/// The representation of a color scale.
-/// The first `ColorStop` (position 0.0) defines the left end color.
-/// The last `ColorStop` (position 1.0) defines the right end color.
-#[derive(Debug, Clone)]
-pub struct ColorScale {
-    color_stops: Vec<ColorStop>,
-}
-
-impl ColorScale {
-    /// Create an empty `ColorScale`.
-    pub fn empty() -> Self {
-        Self {
-            color_stops: Vec::new(),
-        }
-    }
-
-    /// Add a `Color` at the given position.
-    pub fn add_stop(&mut self, color: Color, position: Fraction) -> &mut Self {
-        #![allow(clippy::float_cmp)]
-        let same_position = self
-            .color_stops
-            .iter_mut()
-            .find(|c| position.value() == c.position.value());
-
-        match same_position {
-            Some(color_stop) => color_stop.color = color,
-            None => {
-                let next_index = self
-                    .color_stops
-                    .iter()
-                    .position(|c| position.value() < c.position.value());
-
-                let index = next_index.unwrap_or(self.color_stops.len());
-
-                let color_stop = ColorStop { color, position };
-
-                self.color_stops.insert(index, color_stop);
-            }
-        };
-
-        self
-    }
-
-    /// Get the color at the given position using the mixing function.
-    ///
-    /// Note:
-    /// - No color is returned if position isn't between two color stops or the `ColorScale` is empty.
-    pub fn sample(
-        &self,
-        position: Fraction,
-        mix: &dyn Fn(&Color, &Color, Fraction) -> Color,
-    ) -> Option<Color> {
-        if self.color_stops.len() < 2 {
-            return None;
-        }
-
-        let left_stop = self
-            .color_stops
-            .iter()
-            .rev()
-            .find(|c| position.value() >= c.position.value());
-
-        let right_stop = self
-            .color_stops
-            .iter()
-            .find(|c| position.value() <= c.position.value());
-
-        match (left_stop, right_stop) {
-            (Some(left_stop), Some(right_stop)) => {
-                let diff_color_stops = right_stop.position.value() - left_stop.position.value();
-                let diff_position = position.value() - left_stop.position.value();
-                let local_position = Fraction::from(diff_position / diff_color_stops);
-
-                let color = mix(&left_stop.color, &right_stop.color, local_position);
-
-                Some(color)
-            }
-            _ => None,
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -989,109 +901,6 @@ mod tests {
 
         let mix_in_hwb = |other| input.mix::<HWBA>(&other, Fraction::from(0.5));
         assert_eq!(hue, mix_in_hwb(Color::graytone(0.5)).hue.value());
-    }
-
-    #[test]
-    fn color_scale_add_preserves_ordering() {
-        let mut color_scale = ColorScale::empty();
-
-        color_scale
-            .add_stop(Color::red(), Fraction::from(0.5))
-            .add_stop(Color::gray(), Fraction::from(0.0))
-            .add_stop(Color::blue(), Fraction::from(1.0));
-
-        assert_eq!(color_scale.color_stops.get(0).unwrap().color, Color::gray());
-        assert_eq!(color_scale.color_stops.get(1).unwrap().color, Color::red());
-        assert_eq!(color_scale.color_stops.get(2).unwrap().color, Color::blue());
-    }
-
-    #[test]
-    fn color_scale_empty_sample_none() {
-        let mix = Color::mix::<Lab>;
-
-        let color_scale = ColorScale::empty();
-
-        let color = color_scale.sample(Fraction::from(0.0), &mix);
-
-        assert_eq!(color, None);
-    }
-
-    #[test]
-    fn color_scale_one_color_sample_none() {
-        let mix = Color::mix::<Lab>;
-
-        let mut color_scale = ColorScale::empty();
-
-        color_scale.add_stop(Color::red(), Fraction::from(0.0));
-
-        let color = color_scale.sample(Fraction::from(0.0), &mix);
-
-        assert_eq!(color, None);
-    }
-
-    #[test]
-    fn color_scale_sample_same_position() {
-        let mix = Color::mix::<Lab>;
-
-        let mut color_scale = ColorScale::empty();
-
-        color_scale
-            .add_stop(Color::red(), Fraction::from(0.0))
-            .add_stop(Color::green(), Fraction::from(1.0))
-            .add_stop(Color::blue(), Fraction::from(0.0))
-            .add_stop(Color::white(), Fraction::from(1.0));
-
-        let sample_blue = color_scale.sample(Fraction::from(0.0), &mix).unwrap();
-        let sample_white = color_scale.sample(Fraction::from(1.0), &mix).unwrap();
-
-        assert_eq!(sample_blue, Color::blue());
-        assert_eq!(sample_white, Color::white());
-    }
-
-    #[test]
-    fn color_scale_sample() {
-        let mix = Color::mix::<Lab>;
-
-        let mut color_scale = ColorScale::empty();
-
-        color_scale
-            .add_stop(Color::green(), Fraction::from(1.0))
-            .add_stop(Color::red(), Fraction::from(0.0));
-
-        let sample_red_green = color_scale.sample(Fraction::from(0.5), &mix).unwrap();
-
-        let mix_red_green = mix(&Color::red(), &Color::green(), Fraction::from(0.5));
-
-        assert_eq!(sample_red_green, mix_red_green);
-    }
-
-    #[test]
-    fn color_scale_sample_position() {
-        let mix = Color::mix::<Lab>;
-
-        let mut color_scale = ColorScale::empty();
-
-        color_scale
-            .add_stop(Color::green(), Fraction::from(0.5))
-            .add_stop(Color::red(), Fraction::from(0.0))
-            .add_stop(Color::blue(), Fraction::from(1.0));
-
-        let sample_red = color_scale.sample(Fraction::from(0.0), &mix).unwrap();
-        let sample_green = color_scale.sample(Fraction::from(0.5), &mix).unwrap();
-        let sample_blue = color_scale.sample(Fraction::from(1.0), &mix).unwrap();
-
-        let sample_red_green = color_scale.sample(Fraction::from(0.25), &mix).unwrap();
-        let sample_green_blue = color_scale.sample(Fraction::from(0.75), &mix).unwrap();
-
-        let mix_red_green = mix(&Color::red(), &Color::green(), Fraction::from(0.50));
-        let mix_green_blue = mix(&Color::green(), &Color::blue(), Fraction::from(0.50));
-
-        assert_eq!(sample_red, Color::red());
-        assert_eq!(sample_green, Color::green());
-        assert_eq!(sample_blue, Color::blue());
-
-        assert_eq!(sample_red_green, mix_red_green);
-        assert_eq!(sample_green_blue, mix_green_blue);
     }
 
     #[test]
